# api-docs-executeme.json
{"openapi":"3.0.1","info":{"title":"ExecuteMe API","description":"Backend APIs for broker login, admin monitoring, positions, and async trade execution.","contact":{"name":"ExecuteMe"},"version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Admin Monitoring","description":"Admin APIs for users and broker sessions"},{"name":"Trade Execution","description":"Admin APIs for trade requests and execution jobs"},{"name":"Positions","description":"Admin APIs for broker positions"},{"name":"Broker OAuth","description":"Public broker login callback APIs"}],"paths":{"/admin/trades/execute":{"post":{"tags":["Trade Execution"],"summary":"Submit a trade request","description":"Creates a trade request, creates one execution job per eligible broker session, and dispatches async workers.","operationId":"execute","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TradeExecutionCommand"}}},"required":true},"responses":{"200":{"description":"Trade request accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TradeRequestView"}}}},"401":{"description":"Admin authentication is required"},"400":{"description":"Invalid request or no executable broker sessions found"},"403":{"description":"Admin APIs are disabled"}},"security":[{"basicAuth":[]}]}},"/broker/angel/callback":{"get":{"tags":["Broker OAuth"],"summary":"Angel One OAuth callback","description":"Receives Angel One OAuth tokens, validates the signed state, and stores encrypted broker session tokens.","operationId":"callback","parameters":[{"name":"auth_token","in":"query","description":"Angel One JWT/access token returned by the OAuth flow","required":true,"schema":{"type":"string"}},{"name":"feed_token","in":"query","description":"Angel One feed token returned by the OAuth flow","required":true,"schema":{"type":"string"}},{"name":"refresh_token","in":"query","description":"Optional Angel One refresh token when returned by the OAuth flow","required":false,"schema":{"type":"string"}},{"name":"state","in":"query","description":"Signed ExecuteMe OAuth state","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Angel One account connected","content":{"text/html":{"schema":{"type":"string"}}}},"400":{"description":"Invalid or expired OAuth callback state"}}}},"/admin/trades/{tradeRequestId}":{"get":{"tags":["Trade Execution"],"summary":"Get trade request","description":"Returns a trade request with its execution jobs and reconciles the aggregate request status first.","operationId":"tradeRequest","parameters":[{"name":"tradeRequestId","in":"path","description":"Trade request id","required":true,"schema":{"type":"integer","format":"int64"},"example":1001}],"responses":{"401":{"description":"Admin authentication is required"},"400":{"description":"Trade request was not found"},"403":{"description":"Admin APIs are disabled"},"200":{"description":"Trade request returned","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TradeRequestView"}}}}},"security":[{"basicAuth":[]}]}},"/admin/trades/{tradeRequestId}/summary":{"get":{"tags":["Trade Execution"],"summary":"Get execution summary","description":"Returns aggregate execution counts by lifecycle status for a trade request.","operationId":"summary","parameters":[{"name":"tradeRequestId","in":"path","description":"Trade request id","required":true,"schema":{"type":"integer","format":"int64"},"example":1001}],"responses":{"401":{"description":"Admin authentication is required"},"403":{"description":"Admin APIs are disabled"},"200":{"description":"Execution summary returned","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExecutionSummaryView"}}}}},"security":[{"basicAuth":[]}]}},"/admin/trades/{tradeRequestId}/executions":{"get":{"tags":["Trade Execution"],"summary":"List execution jobs","description":"Returns all user/broker execution jobs for a trade request.","operationId":"executions","parameters":[{"name":"tradeRequestId","in":"path","description":"Trade request id","required":true,"schema":{"type":"integer","format":"int64"},"example":1001}],"responses":{"401":{"description":"Admin authentication is required"},"403":{"description":"Admin APIs are disabled"},"200":{"description":"Execution jobs returned","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TradeExecutionView"}}}}}},"security":[{"basicAuth":[]}]}},"/admin/telegram-users":{"get":{"tags":["Admin Monitoring"],"summary":"List Telegram users","description":"Returns active Telegram users registered through the Telegram bot allowlist/login flow.","operationId":"telegramUsers","responses":{"401":{"description":"Admin authentication is required"},"403":{"description":"Admin APIs are disabled"},"200":{"description":"Telegram users returned","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TelegramUserView"}}}}}},"security":[{"basicAuth":[]}]}},"/admin/positions":{"get":{"tags":["Positions"],"summary":"List broker positions","description":"Returns the current positions visible through active Angel One broker sessions.","operationId":"positions","responses":{"401":{"description":"Admin authentication is required"},"403":{"description":"Admin APIs are disabled"},"200":{"description":"Positions returned","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BrokerPosition"}}}}}},"security":[{"basicAuth":[]}]}},"/admin/broker-sessions":{"get":{"tags":["Admin Monitoring"],"summary":"List broker sessions","description":"Returns broker session metadata for connected Angel One users without exposing encrypted tokens.","operationId":"brokerSessions","responses":{"401":{"description":"Admin authentication is required"},"200":{"description":"Broker sessions returned","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BrokerSessionView"}}}}},"403":{"description":"Admin APIs are disabled"}},"security":[{"basicAuth":[]}]}}},"components":{"schemas":{"TradeExecutionCommand":{"required":["exchange","orderType","productType","symbol","symbolToken","transactionType"],"type":"object","properties":{"symbol":{"type":"string","description":"Broker trading symbol","example":"RELIANCE-EQ"},"symbolToken":{"type":"string","description":"Broker symbol token required by Angel One order APIs","example":"2885"},"exchange":{"type":"string","description":"Exchange code","example":"NSE"},"variety":{"type":"string","description":"Order variety. Defaults to NORMAL when omitted.","example":"NORMAL","enum":["NORMAL","STOPLOSS","AMO","ROBO"]},"transactionType":{"type":"string","description":"BUY or SELL","example":"BUY","enum":["BUY","SELL"]},"orderType":{"type":"string","description":"Order type","example":"MARKET","enum":["MARKET","LIMIT","STOPLOSS_LIMIT","STOPLOSS_MARKET"]},"quantity":{"minimum":1,"type":"integer","description":"Order quantity","format":"int32","example":1},"duration":{"type":"string","description":"Order duration. Defaults to DAY when omitted.","example":"DAY","enum":["DAY","IOC"]},"price":{"type":"number","description":"Order price. Use 0 for MARKET orders.","example":0},"triggerPrice":{"type":"number","description":"Trigger price for stop-loss orders","example":0},"squareOff":{"type":"number","description":"Square-off value for supported order varieties","example":0},"stopLoss":{"type":"number","description":"Stop-loss value for supported order varieties","example":0},"scripConsent":{"type":"string","description":"Optional scrip consent value required by some Angel One flows","example":"yes"},"productType":{"type":"string","description":"Broker product type","example":"INTRADAY","enum":["INTRADAY","DELIVERY","CARRYFORWARD","MARGIN"]},"userIds":{"type":"array","description":"Internal ExecuteMe user ids to target. Empty means all active Angel One sessions.","example":[1,2],"items":{"type":"integer","description":"Internal ExecuteMe user ids to target. Empty means all active Angel One sessions.","format":"int64"}},"createdBy":{"type":"string","description":"Admin/user label that submitted the request","example":"admin"}},"description":"Admin command for creating a trade request and execution jobs."},"TradeExecutionView":{"type":"object","properties":{"id":{"type":"integer","description":"Execution job id","format":"int64","example":5001},"tradeRequestId":{"type":"integer","description":"Parent trade request id","format":"int64","example":1001},"userId":{"type":"integer","description":"Internal ExecuteMe user id","format":"int64","example":7},"brokerSessionId":{"type":"integer","description":"Broker session used by this execution","format":"int64","example":12},"brokerName":{"type":"string","description":"Broker code","example":"ANGEL_ONE"},"brokerClientId":{"type":"string","description":"Broker-side client/account id","example":"A123456"},"status":{"type":"string","description":"Execution lifecycle status","example":"PENDING","enum":["PENDING","PROCESSING","SUCCESS","FAILED","RETRY_PENDING"]},"brokerOrderId":{"type":"string","description":"Broker order id after successful placement","example":"240530000123456"},"failureReason":{"type":"string","description":"Failure reason when execution fails"},"attemptCount":{"type":"integer","description":"Number of worker attempts","format":"int32","example":1},"processingStartedAt":{"type":"string","description":"When processing began","format":"date-time"},"completedAt":{"type":"string","description":"When execution finished","format":"date-time"},"createdAt":{"type":"string","description":"Execution row creation time","format":"date-time"},"updatedAt":{"type":"string","description":"Execution row last update time","format":"date-time"}},"description":"Execution job for one user and one broker session."},"TradeRequestView":{"type":"object","properties":{"id":{"type":"integer","description":"Trade request id","format":"int64","example":1001},"symbol":{"type":"string","description":"Broker trading symbol","example":"RELIANCE-EQ"},"symbolToken":{"type":"string","description":"Broker symbol token","example":"2885"},"exchange":{"type":"string","description":"Exchange code","example":"NSE"},"variety":{"type":"string","description":"Order variety","example":"NORMAL","enum":["NORMAL","STOPLOSS","AMO","ROBO"]},"transactionType":{"type":"string","description":"BUY or SELL","example":"BUY","enum":["BUY","SELL"]},"orderType":{"type":"string","description":"Order type","example":"MARKET","enum":["MARKET","LIMIT","STOPLOSS_LIMIT","STOPLOSS_MARKET"]},"quantity":{"type":"integer","description":"Order quantity","format":"int32","example":1},"duration":{"type":"string","description":"Order duration","example":"DAY","enum":["DAY","IOC"]},"price":{"type":"number","description":"Order price","example":0},"triggerPrice":{"type":"number","description":"Trigger price for stop-loss orders","example":0},"squareOff":{"type":"number","description":"Square-off value","example":0},"stopLoss":{"type":"number","description":"Stop-loss value","example":0},"scripConsent":{"type":"string","description":"Optional scrip consent"},"productType":{"type":"string","description":"Product type","example":"INTRADAY","enum":["INTRADAY","DELIVERY","CARRYFORWARD","MARGIN"]},"status":{"type":"string","description":"Aggregate request status","example":"PROCESSING","enum":["SUBMITTED","PROCESSING","COMPLETED","PARTIALLY_FAILED","FAILED"]},"createdBy":{"type":"string","description":"Admin/user label that submitted the request","example":"admin"},"createdAt":{"type":"string","description":"Request creation time","format":"date-time"},"executions":{"type":"array","description":"Execution jobs generated from this request","items":{"$ref":"#/components/schemas/TradeExecutionView"}}},"description":"Admin trade intent plus generated execution jobs."},"ExecutionSummaryView":{"type":"object","properties":{"pending":{"type":"integer","description":"Pending job count","format":"int64","example":10},"processing":{"type":"integer","description":"Processing job count","format":"int64","example":2},"success":{"type":"integer","description":"Successful job count","format":"int64","example":88},"failed":{"type":"integer","description":"Failed job count","format":"int64","example":5},"retryPending":{"type":"integer","description":"Retry-pending job count","format":"int64","example":1}},"description":"Aggregate execution-job counts by lifecycle status."},"TelegramUserView":{"type":"object","properties":{"id":{"type":"integer","description":"Internal ExecuteMe user id","format":"int64","example":7},"telegramUserId":{"type":"integer","description":"Telegram platform user id","format":"int64","example":987654321},"telegramUsername":{"type":"string","description":"Telegram username","example":"trader_user"},"fullName":{"type":"string","description":"Telegram display name","example":"Trader User"},"active":{"type":"boolean","description":"Whether this user is currently active"},"createdAt":{"type":"string","description":"User creation time","format":"date-time"}},"description":"Telegram user visible to admins."},"BrokerPosition":{"type":"object","properties":{"userId":{"type":"string","description":"Internal user id as a string","example":"7"},"brokerName":{"type":"string","description":"Broker code","example":"ANGEL_ONE","enum":["ANGEL_ONE"]},"symbol":{"type":"string","description":"Trading symbol","example":"RELIANCE-EQ"},"exchange":{"type":"string","description":"Exchange code","example":"NSE"},"quantity":{"type":"integer","description":"Net quantity","format":"int32","example":10},"averagePrice":{"type":"number","description":"Average price","example":2860.5},"lastTradedPrice":{"type":"number","description":"Last traded price","example":2875.2},"pnl":{"type":"number","description":"Current profit/loss","example":147.0}},"description":"Broker position snapshot."},"BrokerSessionView":{"type":"object","properties":{"id":{"type":"integer","description":"Broker session primary key","format":"int64","example":12},"userId":{"type":"integer","description":"Internal ExecuteMe user id","format":"int64","example":7},"brokerName":{"type":"string","description":"Broker code","example":"ANGEL_ONE"},"brokerClientId":{"type":"string","description":"Broker-side client/account id","example":"A123456"},"tokenGeneratedAt":{"type":"string","description":"When the current tokens were generated","format":"date-time"},"tokenExpiresAt":{"type":"string","description":"When the current access token expires","format":"date-time"},"status":{"type":"string","description":"Session lifecycle status","example":"ACTIVE"},"lastLoginAt":{"type":"string","description":"Last successful broker login time","format":"date-time"},"createdAt":{"type":"string","description":"Session row creation time","format":"date-time"},"updatedAt":{"type":"string","description":"Session row last update time","format":"date-time"}},"description":"Broker connection metadata visible to admins. Token values are never exposed."}},"securitySchemes":{"basicAuth":{"type":"http","description":"HTTP Basic authentication using ADMIN_USERNAME and ADMIN_PASSWORD.","scheme":"basic"}}}}